+++ server/src/test/java/org/opensearch/snapshots/SnapshotResiliencyTests.java
@@ -2069,6 +2069,7 @@ public class SnapshotResiliencyTests extends OpenSearchTestCase {
                     new RemoteSegmentStoreDirectoryFactory(() -> repositoriesService, threadPool),
                     repositoriesServiceReference::get,
                     fileCacheCleaner,
+                    null,
                     new RemoteStoreStatsTrackerFactory(clusterService, settings)
                 );
                 final RecoverySettings recoverySettings = new RecoverySettings(settings, clusterSettings);
@@ -2298,7 +2299,8 @@ public class SnapshotResiliencyTests extends OpenSearchTestCase {
                             namedWriteableRegistry,
                             List.of(),
                             client
-                        )
+                        ),
+                        null
                     )
                 );
                 actions.put(